+++ TTauri/GUI/Widgets/ButtonWidget.cpp
@@ -20,7 +20,7 @@ ButtonWidget::ButtonWidget(std::string const label) :
     box.topMargin = 10.0;
 }
 
-void ButtonWidget::pipelineImagePlaceVertices(gsl::span<GUI::PipelineImage::Vertex>& vertices, size_t& offset)
+void ButtonWidget::pipelineImagePlaceVertices(gsl::span<GUI::PipelineImage::Vertex>& vertices, int& offset)
 {
     required_assert(window);
     backingImage.loadOrDraw(*window, box.currentExtent(), [&](auto image) {
